create JWT token and add bcrypt to password and change user model colume

+++ app/controllers/api/events_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Api::EventsController < ApplicationController
+  before_action :user_exists, except: [:index]
+
   def index
     render json: Event.joins(:user).all.map { |events|
       {
@@ -17,7 +19,7 @@ class Api::EventsController < ApplicationController
     event = Event.create(
       title: params[:title],
       name: params[:name],
-      user_id: params[:user_id]
+      user_id: params[:loginUser_id]
     )
     if event.valid?
       render json: { message: 'Successfully create an event' }, status: 200
@@ -33,17 +35,26 @@ class Api::EventsController < ApplicationController
 
   def update
     event = Event.find(params[:id])
+    if event.user_id ===params[:loginUser_id]
     event.update(
       title: params[:title],
       name: params[:name],
     )
     render json: { message: 'Successfully updated event' }, status: 200
-  end
+    else
+      render json: { message: 'Unauthorised author, unable to update event' }, status: 403
+    end
+
+    end
 
   def destroy
     event = Event.find(params[:id])
-    event.destroy
+    if event.user_id ===params[:loginUser_id]
+      event.destroy
     render json: { message: 'Event deleted successfully' }, status: 200
+    else
+      render json: { message: 'unable to delete event' }, status: 403
+    end
   end
 end
